+++ apps/vendor-okx/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @yuants/vendor-okx
 
-This log was last generated on Wed, 24 Sep 2025 03:51:36 GMT and should not be manually modified.
+This log was last generated on Thu, 25 Sep 2025 03:58:11 GMT and should not be manually modified.
+
+## 0.22.12
+Thu, 25 Sep 2025 03:58:11 GMT
+
+### Patches
+
+- fix trade value error
+
+## 0.22.11
+Wed, 24 Sep 2025 13:17:55 GMT
+
+### Patches
+
+- fix okx trade
 
 ## 0.22.10
 Wed, 24 Sep 2025 03:51:36 GMT